requirements.txt
streamlit
pandas
prophet
plotly
google-generativeai
statsmodels
serpapi